+++ rust/op-reth/crates/node/tests/it/main.rs
@@ -8,7 +8,7 @@ mod rpc;
 
 mod custom_genesis;
 
-mod estimate_gas;
+mod estimate_gas_7825;
 
 mod p2p_version;
